PLUGINS, ACTIONS & FILTERS: This code allows you to add a 'plugin' in the /plugins folder which will augment functionality provided by XOT. It allows you to create single file or folder style plugins (similar to Wordpress) which add functionality to various 'hook' points within the page. These hooks are by no means definitive and hopefully we can add lots more, however for filters to work in a better way some code will need re-written in order to pass 'content' through the filter before being used or echo-ed..
The current files should do nothing to an install. The example plugins are commented out - remove the 'REMOVE_THIS' text from the filenames in order to activate. The 'GCU Plugin' demonstrates some potentially useful functionality:

It adds a banner at the top of the page (login and editor)
It changes the <title> text
It changes the "My Projects" text to "John's Projects"
It adds text to Pod 1
It adds a title slide to data.xml on save

The plugin_one plugin simply inserts HTML comments at action hook points throughout the page. Enable this by taking out REMOVE_THIS and then refresh the page and view the source to reveal it's result...

git-svn-id: https://xerteonlinetoolkits.googlecode.com/svn/trunk@709 912cdd6b-5c7d-d5a7-a2ba-d0f0cdb91641