Wait 10ms after keypress to check $("input").val()
When using *Gap Fill* Interactivity: *Fill in Blank*, the last character is sometimes not in `input.value` and therefore the answer is considered `false`. I assume this is why the `setTimeout` is in there anyway. On iOS a delay of 0 milliseconds is to little, so the answer is not handled as correct and stays red after one enters the last correct character. (Also discussed [here](https://stackoverflow.com/questions/28944171/e-keycode-in-ios-safari-on-keyup-event).)
This problem is discussed very well [here](https://stackoverflow.com/questions/8795283/jquery-get-input-value-after-keypress). The first answer suggesting to use `keyup` instead of `keypress` should not be used in my opinion, because also the Ctrl key for examples fires the `keyup`event. This means (after typing in 1 character) one could press the Ctrl key multiple times and Xerte *Hints* would show.

I think the **best** solution would be using `input` event of the `<input>` element as the second answer suggest. The problem is this breaks compatibility with [older browsers](https://www.w3schools.com/jsref/event_oninput.asp) 😞