Fix for differences in uppercase/lowercase of username in some sso installations
 - After user verification, which some systemes allow to be
   case-insensitive, the current user login details are queried
   - If a user exists in the logindetails table, the username of the
     logindetails table is used, and firstname, lastname are reset to what
     comes from sso

The result is that is a user logs in for the first time, a user is created
in Xerte (logindetails) with th ecase that is returned by the sso at that
point in time.

From then on, if the username matched case in a case insensitive way, the
username recorded in Xerte will be used.