Fix issue with determining used file for imageSequence
imegaSequence does not mention all the files it uses in the xml so the media
and quota panel does not correctly show the files in use.
A similar issue exists for import pages. Not all the sequence files are copied,
only the first and last.
Extend XerteXmlInspector to know which files are used and use that in the media
and quota panel and in determining which files are to be copied when a page is
imported