FATAL: prevent cross-origin error in iframe if Xerte itself is in iframe
To be able to track standalone pages there is some cod eto check if parent exists and if a certain attribute is set.
This gives a DOMException and cross-origin error if Xerte itelf is embedded in en iframe, because than parent exists,
but you are not allowed to inspect an attribute.

Solution is to wrap those checks in a try block and catch the exception.