Added AI functionality to several interactivites, added support for using existing context, updated UIs, fixed bug with file uploads.
Added AI support for:
- Thought provoking question
- Sortable grid
- Hangman
- Inventory
- Summary
- Text Correction
- Text Highlight

Updated All AI interactivities UI with:
- Better file upload UI that hides unecessary elements when not selected
- Checkbox to use context of existing learning object

Added function which takes and filters the existing xml of the learning object to then use as part of the context in the thread chain used to communicate with the model.

Fixed bug where dialog would use the wrong delimeter if the user has not enabled the delimeter property.

Fixed a bug where if user selected the file upload checkbox but did not provide a link, the request would not go through. Also added a notification to remind the user.